refactor: clean up replay feature gating
