+++ madara/crates/client/block_production/src/util.rs
@@ -194,30 +194,11 @@ pub(crate) fn create_execution_context(
     previous_l2_gas_price: u128,
     previous_l2_gas_used: u128,
 ) -> anyhow::Result<BlockExecutionContext> {
+    let (block_timestamp, gas_prices) =
+        current_block_timestamp_and_gas_prices(backend, previous_l2_gas_price, previous_l2_gas_used)?;
     #[cfg(feature = "replay")]
-    let (block_timestamp, gas_prices) = if let Some(custom_header) = backend.get_custom_header(block_n) {
-        // Convert Unix timestamp (seconds since Jan 1, 1970) to SystemTime
-        let block_timestamp = UNIX_EPOCH + Duration::from_secs(custom_header.timestamp);
-        let gas_prices = custom_header.gas_prices;
-        (block_timestamp, gas_prices)
-    } else {
-        let l1_gas_quote = backend
-            .get_last_l1_gas_quote()
-            .context("No L1 gas quote available. Ensure that the L1 gas quote is set before calculating gas prices.")?;
-
-        let gas_prices = backend.calculate_gas_prices(&l1_gas_quote, previous_l2_gas_price, previous_l2_gas_used)?;
-        (SystemTime::now(), gas_prices)
-    };
-
-    #[cfg(not(feature = "replay"))]
-    let (block_timestamp, gas_prices) = {
-        let l1_gas_quote = backend
-            .get_last_l1_gas_quote()
-            .context("No L1 gas quote available. Ensure that the L1 gas quote is set before calculating gas prices.")?;
-
-        let gas_prices = backend.calculate_gas_prices(&l1_gas_quote, previous_l2_gas_price, previous_l2_gas_used)?;
-        (SystemTime::now(), gas_prices)
-    };
+    let (block_timestamp, gas_prices) =
+        replay_block_timestamp_and_gas_prices(backend, block_n).unwrap_or((block_timestamp, gas_prices));
 
     Ok(BlockExecutionContext {
         sequencer_address: **backend.chain_config().sequencer_address,
@@ -226,6 +207,31 @@ pub(crate) fn create_execution_context(
         gas_prices,
         l1_da_mode: backend.chain_config().l1_da_mode,
         block_number: block_n,
+    })
+}
+
+fn current_block_timestamp_and_gas_prices(
+    backend: &Arc<MadaraBackend>,
+    previous_l2_gas_price: u128,
+    previous_l2_gas_used: u128,
+) -> anyhow::Result<(SystemTime, GasPrices)> {
+    let l1_gas_quote = backend
+        .get_last_l1_gas_quote()
+        .context("No L1 gas quote available. Ensure that the L1 gas quote is set before calculating gas prices.")?;
+
+    let gas_prices = backend.calculate_gas_prices(&l1_gas_quote, previous_l2_gas_price, previous_l2_gas_used)?;
+    Ok((SystemTime::now(), gas_prices))
+}
+
+#[cfg(feature = "replay")]
+fn replay_block_timestamp_and_gas_prices(
+    backend: &Arc<MadaraBackend>,
+    block_n: u64,
+) -> Option<(SystemTime, GasPrices)> {
+    backend.get_custom_header(block_n).map(|custom_header| {
+        // Convert Unix timestamp (seconds since Jan 1, 1970) to SystemTime
+        let block_timestamp = UNIX_EPOCH + Duration::from_secs(custom_header.timestamp);
+        (block_timestamp, custom_header.gas_prices)
     })
 }
