+++ madara/crates/client/rpc/src/versions/admin/v0_1_0/methods/mod.rs
@@ -2,3 +2,5 @@ pub mod read;
 pub mod services;
 pub mod status;
 pub mod write;
+#[cfg(feature = "replay")]
+mod write_replay;
